final commit

+++ pkproject/app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Models\Post;
 use DB;
 
@@ -13,6 +14,10 @@ class PostsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function __construct(){
+       $this->middleware('auth',['except' => ['index','show']]);
+    }
     public function index()
     {
         //$posts =  Post::all();
@@ -42,11 +47,30 @@ class PostsController extends Controller
         //
         $this -> validate($request,[
           'title' => 'required',
-          'body' => 'required'
+          'body' => 'required',
+          'cover_image' => 'image|nullable|max:1999'
         ]);
+        //handle the file is_uploaded
+        if($request->hasFile('cover_image')){
+          //get filename with extensions
+          $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+          //get just filename
+          $filename = pathinfo($filenameWithExt,PATHINFO_FILENAME);
+          //get just ext
+          $extension = $request->file('cover_image')->getClientOriginalExtension();
+          //filename to store
+          $fileNameToStore = $filename .'_'.time().'.'.$extension;
+          //upload image
+          $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }else{
+          $fileNameToStore = '/public/storage/cover_images/noimage.jpg';
+        }
+
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success','Post created.');
@@ -73,6 +97,9 @@ class PostsController extends Controller
     public function edit($id)
     {
       $post =  Post::find($id);
+      if(auth()->user()->id !== $post->user_id){
+        return redirect('/posts')->with('error','Unauthorize page.');
+      }
       return view ('posts.edit')->with('post', $post);
     }
 
@@ -87,12 +114,29 @@ class PostsController extends Controller
     {
       $this -> validate($request,[
         'title' => 'required',
-        'body' => 'required'
+        'body' => 'required',
+        'cover_image' => 'image|nullable|max:1999'
       ]);
+      //handle the file is_uploaded
+      if($request->hasFile('cover_image')){
+        //get filename with extensions
+        $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+        //get just filename
+        $filename = pathinfo($filenameWithExt,PATHINFO_FILENAME);
+        //get just ext
+        $extension = $request->file('cover_image')->getClientOriginalExtension();
+        //filename to store
+        $fileNameToStore = $filename .'_'.time().'.'.$extension;
+        //upload image
+        $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+      }
 
       $post =  Post::find($id);
       $post->title = $request->input('title');
       $post->body = $request->input('body');
+      if($request->hasFile('cover_image')){
+        $post->cover_image = $fileNameToStore;
+      }
       $post->save();
 
       return redirect('/posts')->with('success','Post updated.');
@@ -108,6 +152,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
       $post =  Post::find($id);
+      if(auth()->user()->id !== $post->user_id){
+        return redirect('/posts')->with('error','Unauthorize page.');
+      }
+      if($post->cover_image != '/public/storage/cover_images/noimage.jpg'){
+        Storage::delete('public/cover_images/'.$post->cover_image);
+      }
       $post->delete();
       return redirect('/posts')->with('success','Post removed');
     }
